chore: add Next.js and update project to SSR
