+++ main/package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "next dev",
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "lint:fix": "eslint ./src --fix",
@@ -16,11 +16,10 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.5.1",
+    "next": "^15.2.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-redux": "^9.2.0",
-    "react-router": "^7.1.5",
-    "react-router-dom": "^7.2.0"
+    "react-redux": "^9.2.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
@@ -28,6 +27,7 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.6.1",
+    "@types/node": "22.13.8",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "@vitejs/plugin-react": "^4.3.4",
