+++ main/pages/_app.tsx
@@ -1,0 +1,12 @@
+import type { AppProps } from 'next/app';
+import '../styles/global.css';
+import { Provider } from 'react-redux';
+import { store } from '../src/store';
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <Provider store={store}>
+      <Component {...pageProps} />
+    </Provider>
+  );
+}
