+++ main/pages/users.tsx
@@ -1,0 +1,8 @@
+const Users = () => {
+  return (
+    <div>
+      <h1>F:FFLLF</h1>
+    </div>
+  );
+};
+export default Users;
