--- main/src/components/Button/Button.css
@@ -1,22 +1,0 @@
-button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  height: 40px;
-  max-width: 80px;
-  flex: 1 0 0;
-  border-radius: 16px;
-  background: var(--text-light);
-  box-shadow: 0px 1px 3px 1px rgba(0, 0, 0, 0.25) inset;
-  border: none;
-  cursor: pointer;
-  color: var(--primary-bg);
-  font-size: 14px;
-  font-weight: bold;
-  text-transform: uppercase;
-}
-
-button:hover {
-  text-decoration: underline;
-}
