+++ main/src/components/Button/Button.tsx
@@ -1,5 +1,5 @@
 import { ButtonProps } from '../../types/types';
-import './Button.css';
+import './Button.module.css';
 
 function Button({ className, onClick, children }: ButtonProps) {
   return (
