+++ main/src/components/CardLoader/CardLoader.tsx
@@ -1,14 +1,14 @@
-import '../Card/Card.css';
-import './CardLoader.css';
-import pokeLoader from '/assets/img/poke-loader.png';
+import '../Card/Card.module.css';
+import './CardLoader.module.css';
+// import pokeLoader from '/assets/img/poke-loader.png';
 
 function CardLoader() {
-  const imageUrl = pokeLoader;
+  // const imageUrl = pokeLoader;
 
   return (
     <div className="card card-loader">
       <span className="card-number">#000</span>
-      <img className="card-img" src={imageUrl} alt="loader" />
+      <img className="card-img" src="{imageUrl}" alt="loader" />
       <div className="card-title">
         <h3>Pokémon Name</h3>
         <p>Abilities: Some abilities</p>
