--- main/src/main.tsx
@@ -1,19 +1,0 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
-import { BrowserRouter } from 'react-router-dom';
-import { Provider } from 'react-redux';
-import { store } from './store/index.ts';
-import { ThemeProvider } from './context/ThemeProvider.tsx';
-
-createRoot(document.getElementById('root') as HTMLElement).render(
-  <StrictMode>
-    <BrowserRouter>
-      <ThemeProvider>
-        <Provider store={store}>
-          <App />
-        </Provider>
-      </ThemeProvider>
-    </BrowserRouter>
-  </StrictMode>
-);
