--- main/src/views/Header/Header.css
@@ -1,22 +1,0 @@
-header {
-  display: flex;
-  padding: 12px 12px 24px 12px;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 16px;
-  align-self: stretch;
-}
-.title {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  align-self: stretch;
-}
-
-span {
-  color: var(--text-light);
-  font-size: 32px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 32px;
-}
