+++ main/src/views/Header/Header.module.css
@@ -1,0 +1,6 @@
+.title {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  align-self: stretch;
+}
