+++ main/styles/global.css
@@ -1,0 +1,164 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+:root {
+  font-family: Poppins;
+
+  --primary-bg: #dc0a2d;
+  --secondary-bg: #fff;
+  --grayscale-bg: #efefef;
+  --border: #ccc;
+  --placeholder-color: #aaa;
+  --text: #666;
+  --text-light: #fff;
+  --text-dark: #1d1d1d;
+  --text-attr: #1d1d1d;
+  --divider: #e0e0e0;
+  --box-shadow-inset: 0px 1px 3px 1px rgba(0, 0, 0, 0.25) inset;
+  --box-shadow-card: 0px 1px 3px 1px rgba(0, 0, 0, 0.2);
+  --box-shadow-strong: 0 4px 8px rgba(0, 0, 0, 0.2);
+  --box-shadow-light: 0 2px 4px rgba(0, 0, 0, 0.2);
+  --btn-disabled-bg: rgba(255, 255, 255, 0.3);
+  --btn-disabled-border: rgba(255, 255, 255, 0.5);
+  --btn-disabled-text: rgba(255, 255, 255, 0.7);
+  --text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.3);
+  --error-bg: #ffe6e6;
+  --error-border: #ff0000;
+  --error-btn: #ff4d4d;
+  --error-btn-hover: #e60000;
+}
+
+[data-theme='dark'] {
+  --primary-bg: #55091b;
+  --secondary-bg: #1e1e1e;
+  --grayscale-bg: #a8a8a8;
+  --border: #666;
+  --placeholder-color: #666;
+  --text: #f5f5f5;
+  --text-light: #e5e5e5;
+  --text-dark: #1d1d1d;
+  --text-attr: #e5e5e5;
+  --divider: #3a3a3a;
+  --box-shadow-inset: 0px 1px 3px 1px rgba(255, 255, 255, 0.25) inset;
+  --box-shadow-card: 0px 1px 3px 1px rgba(255, 255, 255, 0.2);
+  --box-shadow-strong: 0 4px 12px rgba(255, 255, 255, 0.2);
+  --box-shadow-light: 0 2px 6px rgba(255, 255, 255, 0.2);
+  --btn-disabled-bg: rgba(255, 255, 255, 0.1);
+  --btn-disabled-border: rgba(255, 255, 255, 0.2);
+  --btn-disabled-text: rgba(255, 255, 255, 0.4);
+  --text-shadow: 1px 1px 2px rgba(255, 255, 255, 0.3);
+  --error-bg: #400000;
+  --error-border: #55091b;
+  --error-btn: #b30000;
+  --error-btn-hover: #990000;
+}
+
+body {
+  background-color: var(--primary-bg);
+}
+
+a {
+  text-decoration: none;
+  color: inherit;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.container {
+  display: flex;
+  margin: 0 auto;
+  max-width: 1440px;
+  padding: 10px;
+  flex-direction: column;
+  align-items: center;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg) translateY(0);
+  }
+  50% {
+    transform: rotate(180deg) translateY(-10px);
+  }
+  100% {
+    transform: rotate(360deg) translateY(0);
+  }
+}
+
+.broken {
+  display: flex;
+  padding: 10px;
+  max-width: 1440px;
+  max-height: 100vh;
+  width: 100%;
+  height: 100%;
+  flex-direction: column;
+  gap: 20px;
+  align-items: center;
+  justify-content: center;
+  border-radius: 8px;
+  background-color: var(--secondary-bg);
+  box-shadow: var(--box-shadow-inset);
+}
+
+.broken span {
+  color: var(--text-dark);
+  font-size: 24px;
+  font-weight: normal;
+}
+.broken-pokeball {
+  display: flex;
+  width: 400px;
+  height: 400px;
+}
+
+button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  height: 40px;
+  max-width: 80px;
+  flex: 1 0 0;
+  border-radius: 16px;
+  background: var(--text-light);
+  box-shadow: 0px 1px 3px 1px rgba(0, 0, 0, 0.25) inset;
+  border: none;
+  cursor: pointer;
+  color: var(--primary-bg);
+  font-size: 14px;
+  font-weight: bold;
+  text-transform: uppercase;
+}
+
+button:hover {
+  text-decoration: underline;
+}
+
+h3 {
+  color: var(--text-dark);
+  text-transform: capitalize;
+  text-align: center;
+  flex: 1 0 0;
+  margin: 0;
+}
+p {
+  text-align: center;
+  margin: 0;
+}
+
+header {
+  display: flex;
+  padding: 12px 12px 24px 12px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 16px;
+  align-self: stretch;
+}
+
+span {
+  color: var(--text-light);
+  font-size: 32px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 32px;
+}
